Cleanup
A real main in a main.cpp file has been created.
A Makefile has been created.
A LFSR type (in the LFSR_indexes class) is the type definition for all the project.
A test function can check the polynomial coefficients for all the supported sizes.

TODO decide what to do with the unsigned long hardcaded for the states counters.